Adds a doDidChangeQuery option when updating query.
If and only if this option is set to false, the didChangeQuery()
callback will NOT trigger after updating the query. This is useful for
things like autocompletion or when you'd want to update the query text
while bypassing code that recomputes and filters the list items.

This is the only change I needed to make to `SelectListView` to fully
implement autocompletion (see #12) on top of `atom-select-list`.